Split out move processing from mover's logic.
Separate off parsing of moves and basic processing of direction enums
from mover's logic file into its own file.  That way, we can cleanly use
that logic as well for the upcoming PendingMoveProcessor.
