+++ mover/Makefile.am
@@ -17,9 +17,11 @@ libmover_la_LIBADD = \
   $(JSONCPP_LIBS) $(GLOG_LIBS) $(PROTOBUF_LIBS)
 libmover_la_SOURCES = \
   logic.cpp \
+  moves.cpp \
   proto/mover.pb.cc
 noinst_HEADERS = \
   logic.hpp \
+  moves.hpp \
   proto/mover.pb.h
 
 moverd_CXXFLAGS = \
@@ -43,7 +45,8 @@ tests_CXXFLAGS = \
 tests_LDADD = $(builddir)/libmover.la \
   $(JSONCPP_LIBS) $(GLOB_LIBS) $(PROTOBUF_LIBS) $(GTEST_LIBS)
 tests_SOURCES = \
-  logic_tests.cpp
+  logic_tests.cpp \
+  moves_tests.cpp
 
 proto/mover.pb.h proto/mover.pb.cc: $(srcdir)/proto/mover.proto
 	protoc --cpp_out=. "$<"
